FIx issue where inlcuding loader.h would include a ton of junk.
